+++ bm/About/CardViewController.swift
@@ -61,7 +61,7 @@ class CardViewController: UIViewController {
     }
 
     func barCode() -> String? {
-        guard let credentials = Credentials.load(from: .standard) else {
+        guard let credentials = Credentials.sharedCredentials() else {
             return nil
         }