+++ bm/Libraries/LibrariesViewController.swift
@@ -21,11 +21,18 @@ class LibrariesViewController: UIViewController, UITableViewDelegate, UITableVie
     @IBOutlet var tableView: UITableView?
     @IBOutlet var mapView: MKMapView?
     @IBOutlet var showUserLocationButton: UIButton?
+    @IBOutlet var separatorWidth: NSLayoutConstraint?
+    @IBOutlet var separatorHeight: NSLayoutConstraint?
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
         navigationController?.configureCustomAppearance()
+
+        tableView?.tableFooterView = UIView(frame: CGRect.zero)
+
+        separatorWidth?.constant = 1.0 / UIScreen.main.scale
+        separatorHeight?.constant = 1.0 / UIScreen.main.scale
 
         mapView?.setRegion(defaultcoordinateRegion, animated: false)