+++ bm/Login/LoginViewController.swift
@@ -49,15 +49,19 @@ class LoginViewController: UIViewController, UITextFieldDelegate {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
+        #if !targetEnvironment(macCatalyst)
         statusBarBackground?.frame = UIApplication.shared.statusBarFrame
+        #endif
     }
 
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
         super.viewWillTransition(to: size, with: coordinator)
 
+        #if !targetEnvironment(macCatalyst)
         coordinator.animateAlongsideTransition(in: nil, animation: nil) { _ in
             self.statusBarBackground?.frame = UIApplication.shared.statusBarFrame
         }
+        #endif
     }
 
     override var preferredStatusBarStyle: UIStatusBarStyle {
@@ -126,7 +130,7 @@ class LoginViewController: UIViewController, UITextFieldDelegate {
         _ = Authenticate.authenticate(username: subscriberNumber, password: password) { result in
             switch result {
             case .success(let credentials):
-                credentials.save(to: .standard)
+                credentials.save(to: Credentials.defaultKeychain())
                 self.fetchItems(with: credentials)
 
             case .failure(let error):